services/training_keras/requirements.txt
kafka-python
tensorflow
numpy
psutil